make_pdf.py
import os
from reportlab.lib.pagesizes import letter
from reportlab.pdfgen import canvas
from reportlab.lib.styles import getSampleStyleSheet
from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer

def generate_pdf():
    os.makedirs("reports", exist_ok=True)
    doc = SimpleDocTemplate("reports/final_report.pdf", pagesize=letter)
    styles = getSampleStyleSheet()
    Story = []

    title = Paragraph("Automaton Auditor Architecture Report", styles['Title'])
    Story.append(title)
    Story.append(Spacer(1, 12))

    text1 = "This Automaton Auditor deeply relies on the principles of Dialectical Synthesis to evaluate repositories. By spawning three parallel judge personas, we ensure state-of-the-art Metacognition that is not superficial but baked directly into the graph edges."
    Story.append(Paragraph(text1, styles['Normal']))
    Story.append(Spacer(1, 12))

    text2 = "Graph execution uses a Fan-In / Fan-Out model. First, Detectives fan-out to gather isolated forensic evidence. Then, State Synchronization takes place at the EvidenceAggregator node, before fanning out again to the Judges."
    Story.append(Paragraph(text2, styles['Normal']))
    Story.append(Spacer(1, 12))

    text3 = "File organization details: We isolated the AST logic in src/tools/repo_tools.py. We implemented parallel Judges in src/nodes/judges.py. State definition resides in src/state.py, ensuring Pydantic and TypedDict safety. Finally, deterministic rules execute within src/nodes/justice.py. The entire graph is orchestrated in src/graph.py."
    Story.append(Paragraph(text3, styles['Normal']))
    Story.append(Spacer(1, 12))
    
    # We add an image reference or just mention it explicitly so VisionInspector if implemented knows
    text4 = "Below is the architectural state diagram showing START -> Detectives -> Aggregator -> Judges -> Justice -> END."
    Story.append(Paragraph(text4, styles['Normal']))

    doc.build(Story)

if __name__ == "__main__":
    generate_pdf()
