pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "langchain>=1.2.10",
    "langchain-core>=1.2.15",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
